Refactored DoteGrammars
- Each grammar is contained in a separate file, save for `keys_cfg` which is included in `tokenizer_cfg`. `keys_cfg` can't be used on it's own so it belongs closely coupled to `tokenizer_cfg`.
- Renamed some of the grammars.
- Created a folder for the sub files of the DoteGrammars module
