+++ test/token_seq_test.rb
@@ -39,7 +39,7 @@ end
 
 describe Dote::TokenPass::TokenSeq do
   before do
-    @lang = Dote::DoteGrammars.format
+    @lang = Dote::DoteGrammars.display_fmt
     @alternation_rule = @lang.sub_string
     @concatenation_rule = @lang.variable_identifier
     @token_seq = Dote::TokenPass::TokenSeq.new(4) {Dote::LexemeCapture::Token.new}
